Submission

+++ app/Controllers/Backend/SickLeave.php
@@ -6,6 +6,7 @@ use App\Controllers\BaseController;
 use App\Models\M_Absent;
 use App\Models\M_AccessMenu;
 use App\Models\M_Employee;
+use App\Models\M_AbsentDetail;
 use Config\Services;
 
 class SickLeave extends BaseController
@@ -17,6 +18,7 @@ class SickLeave extends BaseController
     {
         $this->request = Services::request();
         $this->model = new M_Absent($this->request);
+        $this->modelDetail = new M_AbsentDetail($this->request);
         $this->entity = new \App\Entities\Absent();
     }
 
@@ -117,7 +119,7 @@ class SickLeave extends BaseController
                 $row[] = $value->division;
                 $row[] = format_dmy($value->submissiondate, '-');
                 $row[] = format_dmy($value->startdate, '-') . " s/d " . format_dmy($value->enddate, '-');
-                $row[] = !is_null($value->receiveddate) ? format_dmy($value->receiveddate, '-') : "";
+                $row[] = !is_null($value->receiveddate) ? format_dmytime($value->receiveddate, '-') : "";
                 $row[] = $value->reason;
                 $row[] = docStatus($value->docstatus);
                 $row[] = $value->createdby;
@@ -143,23 +145,41 @@ class SickLeave extends BaseController
         if ($this->request->getMethod(true) === 'POST') {
             $post = $this->request->getVar();
             $file = $this->request->getFile('image');
+            $file2 = $this->request->getFile('image2');
+            $file3 = $this->request->getFile('image3');
 
             $post["submissiontype"] = $this->Pengajuan_Sakit;
             $post["necessary"] = 'SK';
 
             try {
                 $img_name = "";
+                $img2_name = "";
+                $img3_name = "";
+                $value = "";
 
-                if ($file && $file->isValid()) {
+                if (!empty($post['md_employee_id'])) {
                     $row = $mEmployee->find($post['md_employee_id']);
-
-                    $ext = $file->getClientExtension();
                     $lenPos = strpos($row->getValue(), '-');
                     $value = substr_replace($row->getValue(), "", $lenPos);
-                    $ymd = date('Ymd', strtotime($post['submissiondate']));
+                    $ymd = date('YmdHis');
+                }
 
+                if ($file && $file->isValid()) {
+                    $ext = $file->getClientExtension();
                     $img_name = $this->Pengajuan_Sakit . '_' . $value . '_' . $ymd . '.' . $ext;
                     $post['image'] = $img_name;
+                }
+
+                if ($file2 && $file2->isValid()) {
+                    $ext2 = $file2->getClientExtension();
+                    $img2_name = $this->Pengajuan_Sakit . '_' . $value . '2_' . $ymd . '.' . $ext2;
+                    $post['image2'] = $img2_name;
+                }
+
+                if ($file3 && $file3->isValid()) {
+                    $ext3 = $file3->getClientExtension();
+                    $img3_name = $this->Pengajuan_Sakit . '_' . $value . '3_' . $ymd . '.' . $ext3;
+                    $post['image3'] = $img3_name;
                 }
 
                 if (!$this->validation->run($post, 'sakit')) {
@@ -169,14 +189,32 @@ class SickLeave extends BaseController
 
                     if ($this->isNew()) {
                         uploadFile($file, $path, $img_name);
+
+                        if ($file2 && $file2->isValid())
+                            uploadFile($file2, $path, $img2_name);
+
+                        if ($file3 && $file3->isValid())
+                            uploadFile($file3, $path, $img3_name);
                     } else {
                         $row = $this->model->find($this->getID());
 
-                        if (!empty($row->getImage()) && $post['image'] !== $row->getImage()) {
-                            if (file_exists($path . $row->getImage())) {
+                        if (!empty($post['image']) && !empty($row->getImage()) && $post['image'] !== $row->getImage()) {
+                            if (file_exists($path . $row->getImage()))
                                 unlink($path . $row->getImage());
-                                $file->move($path);
-                            }
+
+                            uploadFile($file, $path, $img_name);
+                        }
+
+                        if (empty($post['image2']) && !empty($row->getImage2()) && file_exists($path . $row->getImage2())) {
+                            unlink($path . $row->getImage2());
+                        } else {
+                            uploadFile($file2, $path, $img2_name);
+                        }
+
+                        if (empty($post['image3']) && !empty($row->getImage3()) && file_exists($path . $row->getImage3())) {
+                            unlink($path . $row->getImage3());
+                        } else {
+                            uploadFile($file3, $path, $img3_name);
                         }
                     }
 
@@ -184,7 +222,7 @@ class SickLeave extends BaseController
 
                     if ($this->isNew()) {
                         $this->entity->setDocStatus($this->DOCSTATUS_Drafted);
-                        $docNo = $this->model->getInvNumber("submissiontype", $this->Pengajuan_Sakit, $post["necessary"]);
+                        $docNo = $this->model->getInvNumber("submissiontype", $this->Pengajuan_Sakit, $post);
                         $this->entity->setDocumentNo($docNo);
                     }
 
@@ -205,18 +243,39 @@ class SickLeave extends BaseController
         if ($this->request->isAJAX()) {
             try {
                 $list = $this->model->where($this->model->primaryKey, $id)->findAll();
+                $detail = $this->modelDetail->where($this->model->primaryKey, $id)->findAll();
                 $rowEmp = $mEmployee->where($mEmployee->primaryKey, $list[0]->getEmployeeId())->first();
 
-                $path = 'uploads/' . $this->PATH_Pengajuan . '/';
-                $list[0]->setImage($path . $list[0]->getImage());
+                $path = $this->PATH_UPLOAD . $this->PATH_Pengajuan . '/';
+
+                if (file_exists($path . $list[0]->getImage())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage($path . $list[0]->getImage());
+                } else {
+                    $list[0]->setImage(null);
+                }
+
+                if (!empty($list[0]->getImage2()) && file_exists($path . $list[0]->getImage2())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage2($path . $list[0]->getImage2());
+                } else {
+                    $list[0]->setImage2(null);
+                }
+
+                if (!empty($list[0]->getImage3()) && file_exists($path . $list[0]->getImage3())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage3($path . $list[0]->getImage3());
+                } else {
+                    $list[0]->setImage3(null);
+                }
 
                 $list = $this->field->setDataSelect($mEmployee->table, $list, $mEmployee->primaryKey, $rowEmp->getEmployeeId(), $rowEmp->getValue());
-
-                $title = $list[0]->getDocumentNo() . "_" . $rowEmp->getFullName();
 
                 //* Need to set data into date field in form
                 $list[0]->setStartDate(format_dmy($list[0]->startdate, "-"));
                 $list[0]->setEndDate(format_dmy($list[0]->enddate, "-"));
+
+                $title = $list[0]->getDocumentNo() . "_" . $rowEmp->getFullName();
 
                 $fieldHeader = new \App\Entities\Table();
                 $fieldHeader->setTitle($title);
@@ -224,7 +283,8 @@ class SickLeave extends BaseController
                 $fieldHeader->setList($list);
 
                 $result = [
-                    'header'    => $this->field->store($fieldHeader)
+                    'header'    => $this->field->store($fieldHeader),
+                    'line'      => $this->tableLine('edit', $detail)
                 ];
 
                 $response = message('success', true, $result);
@@ -253,7 +313,6 @@ class SickLeave extends BaseController
     public function processIt()
     {
         $cWfs = new WScenario();
-
         if ($this->request->isAJAX()) {
             $post = $this->request->getVar();
 
@@ -269,7 +328,7 @@ class SickLeave extends BaseController
                         $response = message('error', true, 'Silahkan refresh terlebih dahulu');
                     } else if ($_DocAction === $this->DOCSTATUS_Completed) {
                         $this->message = $cWfs->setScenario($this->entity, $this->model, $this->modelDetail, $_ID, $_DocAction, $menu, $this->session);
-                        $response = message('success', true, $this->message);
+                        $response = message('success', true, true);
                     } else if ($_DocAction === $this->DOCSTATUS_Unlock) {
                         $this->entity->setDocStatus($this->DOCSTATUS_Drafted);
                         $response = $this->save();
@@ -288,5 +347,32 @@ class SickLeave extends BaseController
 
             return $this->response->setJSON($response);
         }
+    }
+
+    public function tableLine($set = null, $detail = [])
+    {
+        $table = [];
+
+
+        //? Update
+        if (!empty($set) && count($detail) > 0) {
+            foreach ($detail as $row) :
+                if (!empty($row->ref_absent_detail_id)) {
+                    $line = $this->modelDetail->getDetail('trx_absent_detail_id', $row->ref_absent_detail_id)->getRow();
+                    $doc = $line->documentno;
+                } else {
+                    $doc = "";
+                }
+
+                $table[] = [
+                    $row->lineno,
+                    format_dmy($row->date, '-'),
+                    $doc,
+                    statusRealize($row->isagree)
+                ];
+            endforeach;
+        }
+
+        return json_encode($table);
     }
 }
